fixed search range

+++ BLL/SearchProductLogic.cs
@@ -12,12 +12,12 @@ namespace BLL
                 product.ProductName.Contains(searchValue, StringComparison.InvariantCultureIgnoreCase)).ToList();
         }
 
-        public List<Product> SearchByCalories(List<Product> productList, int searchValue)
+        public List<Product> SearchByCalories(List<Product> productList, double searchValue)
         {
             return productList.Where(product => product.Energy == searchValue).ToList();
         }
 
-        public List<Product> SearchByCalories(List<Product> productList, int minValue, int maxValue)
+        public List<Product> SearchByCalories(List<Product> productList, double minValue, double maxValue)
         {
             return productList.Where(product => product.Energy >= minValue && product.Energy <= maxValue).ToList();
         }
